refactor: 디렉토리 구성, utils 분리
