+++ src/utils/constants.js
@@ -1,9 +1,9 @@
-const MESSAGE = {
+const ABOUT = {
   INPUT_NAME: '경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)\n',
   ASK_LAP: '시도할 횟수는 몇 회인가요?\n',
-  RESULT: '\n실행 결과',
+  RESULT_IS: '\n실행 결과',
   NO_WINNER: '최종 우승자 : 없음',
   FINAL_WINNER: '최종 우승자 : ',
 };
 
-export default MESSAGE;
+export default ABOUT;
